:boom: fix: Trocando o Id do Repositório para Long

+++ src/main/java/biovitta/com/clinics/repositories/PacienteRepositorio.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PacienteRepositorio extends JpaRepository<Paciente, Integer> {
+public interface PacienteRepositorio extends JpaRepository<Paciente, Long> {
 
     boolean existsByEmail(String email);
 }
