+++ src/main/java/biovitta/com/clinics/repositories/UsuarioRepositorio.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepositorio extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepositorio extends JpaRepository<Usuario, Long> {
     UserDetails findByUsuario(String usuario);
 
 }
